Hinge keyframes with rotate+rotateX/Y not tweened
Reproduces in Chrome 60.0.3112.101.  If 0% keyframe's transform contains "rotateX" or "rotateY" and 100% keyframe's transform contains "rotate" (or vice-versa when $turn-origin: from-front) the animation is not tweened properly by the browser and will not render any of the intermediate frames.

Fix is to keep the transforms consistent between start and end (i.e. don't use rotate(0deg), use rotateX(0deg) or rotateY(0deg) depending on $from).